+++ CHANGES.rst
@@ -3,9 +3,9 @@ Changes
 
 0.1.2 -- 2016-08-dd
 -------------------
-* Backwards incompatible change: Removed "eps", "svg", "png", "pdf", and "txt"
-  methods from QRCode. Use QRCode.save. This API change happens without any
-  warning or deprecation hint since Segno is not yet stable.
+* Backwards incompatible change: Deprecated "eps", "svg", "png", "pdf", and
+  "txt" methods from QRCode. Use QRCode.save.
+  Methods will be removed in 0.1.3
 
 
 0.1.1 -- 2016-08-14